Move link component to final destination
