+++ packages/app/src/features/compliance/components/TermsOfService.tsx
@@ -2,8 +2,8 @@ import React, { useState } from 'react'
 
 import { MultiPanelDialog } from '@/features/dialogs/common/components/MultiPanelDialog'
 import { assets } from '@/ui/assets'
+import { Link } from '@/ui/atoms/link/Link'
 import { Button } from '@/ui/atoms/new/button/Button'
-import { Link } from '@/ui/atoms/new/link/Link'
 import { ScrollArea } from '@/ui/atoms/scroll-area/ScrollArea'
 import { Tooltip, TooltipContent, TooltipTrigger } from '@/ui/atoms/tooltip/Tooltip'
 import { links } from '@/ui/constants/links'
