+++ packages/app/src/features/dialogs/savings/migrate/upgrade/views/UpgradeView.tsx
@@ -10,7 +10,7 @@ import { AssetInputSchema } from '@/features/dialogs/common/logic/form'
 import { FormFieldsForDialog, PageStatus } from '@/features/dialogs/common/types'
 import { assets } from '@/ui/assets'
 import { DialogTitle } from '@/ui/atoms/dialog/Dialog'
-import { Link } from '@/ui/atoms/new/link/Link'
+import { Link } from '@/ui/atoms/link/Link'
 import { links } from '@/ui/constants/links'
 import { Info } from '@/ui/molecules/info/Info'
 import { UseFormReturn } from 'react-hook-form'
