+++ packages/app/src/features/easy-borrow/components/form/Deposits.tsx
@@ -1,8 +1,8 @@
 import { paths } from '@/config/paths'
 import { TokenWithBalance } from '@/domain/common/types'
 import { TokenSymbol } from '@/domain/types/TokenSymbol'
+import { Link } from '@/ui/atoms/link/Link'
 import { Button, ButtonIcon } from '@/ui/atoms/new/button/Button'
-import { Link } from '@/ui/atoms/new/link/Link'
 import { Panel } from '@/ui/atoms/new/panel/Panel'
 import { Info } from '@/ui/molecules/info/Info'
 import { AssetInput } from '@/ui/organisms/new/asset-input/AssetInput'
