+++ packages/app/src/features/farm-details/components/farm-info-panel/inactive/InactiveFarmInfoPanel.tsx
@@ -3,8 +3,8 @@ import { farmAddresses } from '@/config/chain/constants'
 import { formatPercentage } from '@/domain/common/format'
 import { AssetsGroup, Farm } from '@/domain/farms/types'
 import { USD_MOCK_TOKEN } from '@/domain/types/Token'
+import { Link } from '@/ui/atoms/link/Link'
 import { Button } from '@/ui/atoms/new/button/Button'
-import { Link } from '@/ui/atoms/new/link/Link'
 import { Panel } from '@/ui/atoms/new/panel/Panel'
 import { links } from '@/ui/constants/links'
 import { testIds } from '@/ui/utils/testIds'
