+++ packages/app/src/features/navbar/components/airdrop-badge/AirdropDetails.tsx
@@ -1,7 +1,7 @@
 import { SPK_MOCK_TOKEN } from '@/domain/types/Token'
 import { assets } from '@/ui/assets'
 import BoxArrowTopRight from '@/ui/assets/box-arrow-top-right.svg?react'
-import { Link } from '@/ui/atoms/new/link/Link'
+import { Link } from '@/ui/atoms/link/Link'
 import { Skeleton } from '@/ui/atoms/skeleton/Skeleton'
 import { links } from '@/ui/constants/links'
 import { NormalizedUnitNumber } from '@marsfoundation/common-universal'
