+++ packages/app/src/features/savings/components/welcome-dialog/WelcomeDialog.tsx
@@ -2,8 +2,8 @@ import { formatPercentage } from '@/domain/common/format'
 import { assets } from '@/ui/assets'
 import { Dialog, DialogContent } from '@/ui/atoms/dialog/Dialog'
 import { KeyPoints } from '@/ui/atoms/key-points/KeyPoints'
+import { Link } from '@/ui/atoms/link/Link'
 import { Button } from '@/ui/atoms/new/button/Button'
-import { Link } from '@/ui/atoms/new/link/Link'
 import { links } from '@/ui/constants/links'
 import { Percentage } from '@marsfoundation/common-universal'
