+++ packages/app/src/features/topbar/components/topbar-menu/components/TopbarMenuAirdropItem.tsx
@@ -2,7 +2,7 @@ import { SPK_MOCK_TOKEN } from '@/domain/types/Token'
 import { useGrowingAirdropAmount } from '@/features/topbar/logic/use-airdrop-info/useGrowingAirdropAmount'
 import { assets } from '@/ui/assets'
 import { Dialog, DialogContent, DialogTitle, DialogTrigger } from '@/ui/atoms/dialog/Dialog'
-import { Link } from '@/ui/atoms/new/link/Link'
+import { Link } from '@/ui/atoms/link/Link'
 import { MenuItem, MenuItemIcon } from '@/ui/atoms/new/menu-item/MenuItem'
 import { Skeleton } from '@/ui/atoms/skeleton/Skeleton'
 import { links } from '@/ui/constants/links'
